processed comments from pull request
