+++ app/src/main/java/com/digitalvotingpass/digitalvotingpass/PassportConnection.java
@@ -23,6 +23,7 @@ public class PassportConnection {
     /**
      * Opens a connection with the ID by doing BAC
      * Uses hardcoded parameters for now
+     *
      * @param tag - NFC tag that started this activity (ID NFC tag)
      * @return PassportService - passportservice that has an open connection with the ID
      */
@@ -42,10 +43,12 @@ public class PassportConnection {
                 public String getDocumentNumber() {
                     return "NP0811B03";
                 }
+
                 @Override
                 public String getDateOfBirth() {
                     return "940610";
                 }
+
                 @Override
                 public String getDateOfExpiry() {
                     return "180624";
@@ -63,6 +66,7 @@ public class PassportConnection {
 
     /**
      * Retrieves the public key used for Active Authentication from datagroup 15.
+     *
      * @return Publickey - returns the publickey used for AA
      */
     public PublicKey getAAPublicKey(PassportService ps) {
@@ -71,29 +75,6 @@ public class PassportConnection {
             is15 = ps.getInputStream(PassportService.EF_DG15);
             DG15File dg15 = new DG15File(is15);
             return dg15.getPublicKey();
-        } catch (Exception ex){
-            ex.printStackTrace();
-        } finally {
-            try {
-                is15.close();
-            } catch (Exception ex) {
-                ex.printStackTrace();
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Signs 8 bytes by the passport using the AA functionality.
-     * @return byte[] - signed byte array
-     */
-    public byte[] signData(PassportService ps, PublicKey pubk) {
-        InputStream is15 = null;
-        try {
-            is15 = ps.getInputStream(PassportService.EF_DG15);
-            // test 8 byte string for testing purposes
-            byte[] data = hexStringToByteArray("0a1b3c4d5e6faabb");
-            return ps.doAA(pubk, null, null, data);
         } catch (Exception ex) {
             ex.printStackTrace();
         } finally {
@@ -107,8 +88,34 @@ public class PassportConnection {
     }
 
     /**
-     *  Get the BSN from datagroup1 to confirm the ID was scanned correctly
-     *  This is for testing purposes
+     * Signs 8 bytes by the passport using the AA functionality.
+     *
+     * @return byte[] - signed byte array
+     */
+    public byte[] signData(PassportService ps) {
+        InputStream is15 = null;
+        try {
+            is15 = ps.getInputStream(PassportService.EF_DG15);
+            // test 8 byte string for testing purposes
+            byte[] data = Util.hexStringToByteArray("0a1b3c4d5e6faabb");
+            // doAA of JMRTD library only returns signed data, and does not have the AA functionality yet
+            // there is no need for sending public key information with the method.
+            return ps.doAA(null, null, null, data);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        } finally {
+            try {
+                is15.close();
+            } catch (Exception ex) {
+                ex.printStackTrace();
+            }
+        }
+        return null;
+    }
+
+    /**
+     * Get the BSN from datagroup1 to confirm the ID was scanned correctly
+     * This is for testing purposes
      */
     public String getBSN(PassportService ps) {
         InputStream is = null;
@@ -116,7 +123,7 @@ public class PassportConnection {
             is = ps.getInputStream(PassportService.EF_DG1);
             DG1File dg1 = (DG1File) LDSFileUtil.getLDSFile(PassportService.EF_DG1, is);
             return dg1.getMRZInfo().getPersonalNumber();
-        } catch (Exception ex){
+        } catch (Exception ex) {
             ex.printStackTrace();
         } finally {
             try {
@@ -126,41 +133,5 @@ public class PassportConnection {
             }
         }
         return null;
-    }
-
-    /**
-     * Method for converting a hexString to a byte array.
-     * This method is used for signing transaction hashes (which are in hex).
-     */
-    public static byte[] hexStringToByteArray(String hStr) {
-        if(hStr != null) {
-            int len = hStr.length();
-            byte[] data = new byte[len / 2];
-            for (int i = 0; i < len; i += 2) {
-                data[i / 2] = (byte) ((Character.digit(hStr.charAt(i), 16) << 4)
-                        + Character.digit(hStr.charAt(i + 1), 16));
-            }
-            return data;
-        }
-        return new byte[0];
-    }
-
-    /**
-     * Method for converting a byte array to a hexString.
-     * This method is used for converting a signed 8-byte array back to a hashString in order to
-     * display it readable.
-     */
-    public static String byteArrayToHexString(byte[] bArray) {
-        if (bArray != null) {
-            final char[] hexArray = "0123456789ABCDEF".toCharArray();
-            char[] hexChars = new char[bArray.length * 2];
-            for (int j = 0; j < bArray.length; j++) {
-                int v = bArray[j] & 0xFF;
-                hexChars[j * 2] = hexArray[v >>> 4];
-                hexChars[j * 2 + 1] = hexArray[v & 0x0F];
-            }
-            return new String(hexChars);
-        }
-        return "";
     }
 }
