+++ app/src/main/java/com/digitalvotingpass/digitalvotingpass/Util.java
@@ -1,0 +1,39 @@
+package com.digitalvotingpass.digitalvotingpass;
+
+public class Util {
+    /**
+     * Method for converting a hexString to a byte array.
+     * This method is used for signing transaction hashes (which are in hex).
+     */
+    public static byte[] hexStringToByteArray(String hStr) {
+        if(hStr != null) {
+            int len = hStr.length();
+            byte[] data = new byte[len / 2];
+            for (int i = 0; i < len; i += 2) {
+                data[i / 2] = (byte) ((Character.digit(hStr.charAt(i), 16) << 4)
+                        + Character.digit(hStr.charAt(i + 1), 16));
+            }
+            return data;
+        }
+        return new byte[0];
+    }
+
+    /**
+     * Method for converting a byte array to a hexString.
+     * This method is used for converting a signed 8-byte array back to a hashString in order to
+     * display it readable.
+     */
+    public static String byteArrayToHexString(byte[] bArray) {
+        if (bArray != null) {
+            final char[] hexArray = "0123456789ABCDEF".toCharArray();
+            char[] hexChars = new char[bArray.length * 2];
+            for (int j = 0; j < bArray.length; j++) {
+                int v = bArray[j] & 0xFF;
+                hexChars[j * 2] = hexArray[v >>> 4];
+                hexChars[j * 2 + 1] = hexArray[v & 0x0F];
+            }
+            return new String(hexChars);
+        }
+        return "";
+    }
+}
