+++ app/src/main/res/layout/activity_passport_con.xml
@@ -1,28 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.digitalvotingpass.digitalvotingpass.PassportConActivity">
+    android:orientation="vertical">
 
     <TextView
-        android:id="@+id/textView"
-        android:layout_width="349dp"
+        android:id="@+id/notice"
+        android:layout_width="match_parent"
         android:layout_height="169dp"
-        android:scrollHorizontally="false"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintHorizontal_bias="0.142"
-        app:layout_constraintVertical_bias="0.052" />
+        android:scrollHorizontally="true" />
 
     <TextView
         android:id="@+id/signedData"
-        android:layout_width="344dp"
-        android:layout_height="221dp"
-        tools:layout_editor_absoluteY="264dp"
-        tools:layout_editor_absoluteX="20dp" />
+        android:layout_width="match_parent"
+        android:layout_height="221dp" />
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
